improve bomb physics support - move to its own class - support multiple explosions - improve physic debug draw - convert all objects to do transforms with floats (greatly improves display of jagged weirdness too)

+++ ninja-engine/camera.cpp
@@ -39,7 +39,7 @@ void Camera::ClampToLevelBounds() {
 		pos.y = levelHeight - sh;
 }
 
-void Camera::TransformWorldToView(int & x, int & y)
+void Camera::TransformWorldToView(float & x, float & y)
 {
 	x -= pos.x + camera_offset.x;
 	y -= pos.y + camera_offset.y;
